Make compilable - add in some unit information, commands.

+++ Emulator/src/emulator/core/Memory.java
@@ -1,10 +1,9 @@
 package emulator.core;
 
-
 public class Memory {
 
 	// byte array representing the memory;
-	public byte[] m;
-	
-	
+	// Using int because byte is signed in java, causing lots of probs!
+	public int[] m;
+
 }
